refactor: getFormattedTimeRanges 함수 lib 폴더로 분리
